Create the page components directly into the iframe
Change the way the page components are created and instead of creating
them and attaching them to the iframe body after they are created,
directly create them from inside the iframe. This allows for proper hot
module reloading because the component will re-generate it's styles in
the correct root and the reference to the global variables like `window`
or `document` will point to the correct instance.